+++ CHANGELOG.md
@@ -54,6 +54,7 @@
       in situations where the write payload isn't received as a unit by the terminal (typically over `ssh`). Respectable terminals will simply ignore these codes if they do not support
       the feature, but just in case, it can be disabled with the `screen-update-sync` variable.
     - More default command completions.
+    - Added SIGABRT to the list of fatal signals that print a backtrace.
 
 ## 1403 - 2022-5-4
 ### Fixed